Added a Commandline script "ics.py" and acceptance tests.
"ics.py" contains code that will be useful for commandline operations.
In addition, added accTests that contains a sample acceptance test.  Test
data is usually associated with an acceptance test, so mdata was moved to
this directory. Finally, moved icse output to a "output" directory, which is
ignored by git. Also, added some logging to "icse.py".